refactor: cell guide sitemap
