+++ frontend/src/pages/cellguide-sitemap.xml/index.tsx
@@ -1,28 +1,8 @@
 import { GetServerSideProps } from "next";
 import { getServerSideSitemap, ISitemapField } from "next-sitemap";
+import { convertTimestamp } from "src/common/sitemaps/utils";
+import { CELLGUIDE_DATA_URL } from "src/configs/configs";
 import xml2js from "xml2js";
-
-function convertTimestamp(originalTimestamp: string) {
-  const date = new Date(originalTimestamp);
-  const year = date.getFullYear();
-  const month = date.getMonth() + 1;
-  const day = date.getDate();
-  const hour = date.getHours();
-  const minute = date.getMinutes();
-  const second = date.getSeconds();
-  const offset = date.getTimezoneOffset();
-  const offsetHours = Math.abs(Math.floor(offset / 60));
-  const offsetMinutes = Math.abs(offset % 60);
-  const timezoneOffset = `${offset < 0 ? "+" : "-"}${offsetHours
-    .toString()
-    .padStart(2, "0")}:${offsetMinutes.toString().padStart(2, "0")}`;
-
-  return `${year}-${month.toString().padStart(2, "0")}-${day
-    .toString()
-    .padStart(2, "0")}T${hour.toString().padStart(2, "0")}:${minute
-    .toString()
-    .padStart(2, "0")}:${second.toString().padStart(2, "0")}${timezoneOffset}`;
-}
 
 interface CellGuideData {
   ListBucketResult: {
@@ -34,9 +14,7 @@ interface CellGuideData {
 }
 
 export const getServerSideProps: GetServerSideProps = async (context) => {
-  const response = await fetch(
-    "https://cellguide.cellxgene.dev.single-cell.czi.technology"
-  );
+  const response = await fetch(CELLGUIDE_DATA_URL);
 
   const xmlData = await response.text();
